Sort games by exact percents but still display rounded value

+++ app/src/main/java/com/jc/steamachievementschecker/data/db/GameInfoDbEntity.kt
@@ -7,7 +7,7 @@ import androidx.room.PrimaryKey
 data class GameInfoDbEntity(
     @PrimaryKey val id: Int,
     val name: String,
-    val achievementsPercentage: Int,
+    val achievementsPercentage: Double,
     val achievementsUnlocked: Int = 0,
     val achievementsTotal: Int = 0,
     val hasNoAchievements: Boolean = false
